feat: added level style scene transitions
Added a level style scene transition that uses a new Phaser Scene.
This works by having a top level scene that will have a black rectangle
take up the whole screen, and we apply a mask to the shape. For the mask,
we invert the alpha that way only circle part will be visible in our game
and as this grows/shrinks, we show more or less of the game.

This results in a level transition effect similar to the classic link
to the past style type transition.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>