feat: add support for item usage
added support for a new item component and use item input key to support having a sword and item equipped at the same time. the new item component is tied to items like the bow, were the player can swap this item out at a later time. the item component is another instance of the weapon component.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>